refactor(messages): 拆分消息渲染大文件
