+++ src/features/messages/components/messagesConstants.ts
@@ -1,0 +1,6 @@
+export const MESSAGE_JUMP_EVENT_NAME = "ccgui:jump-to-message";
+export const ASSISTANT_FINALIZING_LIVE_WINDOW_MS = 320;
+export const CODEX_FINALIZING_LIVE_WINDOW_MS = 6_000;
+export const VISIBLE_TEXT_REPORT_MIN_INTERVAL_MS = 120;
+export const VISIBLE_TEXT_REPORT_MIN_GROWTH_CHARS = 160;
+export const VISIBLE_TEXT_REPORT_EAGER_PREFIX_CHARS = 512;
